requirements.txt
pandas~=2.2.3
numpy~=2.2.1
streamlit~=1.41.1
pillow~=11.0.0